rename src/renderer/partials
