+++ src/renderer/pages/vpn.vue
@@ -72,7 +72,7 @@ import type from '../store/types'
 import { mapGetters, mapMutations } from 'vuex'
 import StatsDisplay from '../components/stats-display'
 import ConnectionButton from '../components/connection-button'
-import AppError from '../partials/AppError'
+import AppError from '../partials/app-error'
 import config from '../config'
 import { ActionLooperConfig } from '../store/modules/connection'
 import FavoriteButton from '../components/favorite-button'
