+++ src/renderer/partials/app-error.vue
