+++ src/renderer/partials/app-nav.vue
