+++ src/renderer/partials/app-visual.vue
